+++ auth0__change_password__POST__fetch/index.js
@@ -1,5 +1,5 @@
 /// <reference types="ctx-core" />
-import { fetch__template_pair__new } from '@ctx-core/fetch-undici'
+import { fetch__template_pair__new } from 'ctx-core/fetch'
 import { auth0__authorization__verify } from '../auth0__authorization__verify/index.js'
 import { auth0__token$_ } from '../auth0__token/index.js'
 export const [