+++ auth0__email/index.js
@@ -1,5 +1,4 @@
-import { be_ } from 'ctx-core/be'
-import { be_memo_pair_, memo_, nullish__none_ } from 'ctx-core/rmemo'
+import { be_memo_pair_, nullish__none_ } from 'ctx-core/rmemo'
 import { auth0__opened__set } from '../auth0__opened/index.js'
 import { auth0__userinfo_ } from '../auth0__userinfo/index.js'
 export const [