+++ auth0__jwt__expiration__validate/index.test.ts
@@ -1,6 +1,6 @@
+import FakeTimers, { InstalledClock } from '@sinonjs/fake-timers'
 import { btoa } from 'ctx-core/btoa'
 import { bad_credentials_error_ } from 'ctx-core/error'
-import FakeTimers, { InstalledClock } from '@sinonjs/fake-timers'
 import { test } from 'uvu'
 import { throws } from 'uvu/assert'
 import { auth0__jwt__expiration__validate } from '../index.js'