+++ auth0__passwordless_start__POST__fetch/index.js
@@ -1,4 +1,4 @@
-import { fetch__template_pair__new } from '@ctx-core/fetch-undici'
+import { fetch__template_pair__new } from 'ctx-core/fetch'
 import { AUTH0_DOMAIN_ } from '../AUTH0_DOMAIN/index.js'
 /** @typedef {import('auth0-js').Auth0Error} */
 /** @typedef {import('./index.d.ts').auth0__passwordless_start__POST__fetch2__body_T} */