+++ jwks__json__GET__fetch/index.d.ts
@@ -1,4 +1,4 @@
-import type { fetch__template_pair_T } from '@ctx-core/fetch'
+import type { fetch__template_pair_T } from 'ctx-core/fetch'
 import type { Auth0Error } from 'auth0-js'
 import type { Ctx } from 'ctx-core/be'
 export declare const [