+++ jwks__json__GET__fetch/index.js
@@ -1,4 +1,4 @@
-import { fetch__template_pair__new } from '@ctx-core/fetch'
+import { fetch__template_pair__new } from 'ctx-core/fetch'
 import { AUTH0_DOMAIN_ } from '../AUTH0_DOMAIN/index.js'
 /** @typedef {import('./index.d.ts').jwks_json_T} */
 export const [